+++ src/oemof/tabular/facades/shortage.py
@@ -1,0 +1,8 @@
+from oemof.tabular.facades import Dispatchable
+
+
+class Shortage(Dispatchable):
+    """ """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
